Update Voice-and-tone.md
The hyperlink to the 18F Content Guide was broken due on Line 28 to a space between the link and the alt-text. This would make it so that the hyperlink shows up on GitHub but not on the final rendering of the website. Removing the space should fix the problem. Thanks!